feat: implement dropList builtin and saturating CostingInteger arithmetic
Add dropList builtin from Plutus 1.53 specification:
- Add DropList to DefaultFun with opcode 88
- Implement runtime semantics in Builtin.scala
- Add DropListCostingFun with IntegerCostedLiterally for cost calculation
- Add dropList to SIRBuiltins for compiler support
- Add cost model parameters to builtinCostModelC.json
- Enable 16 dropList conformance tests (all passing)

Implement saturating arithmetic for CostingInteger:
- Make CostingInteger an opaque type wrapping Long
- Add satPlus, satMinus, satMul functions that saturate to Long.MaxValue/MinValue on overflow instead of wrapping
- Update CountingBudgetSpender to use CostingInteger for accumulation
- Update ExUnits + and |+| operators to use saturating addition
- This matches Plutus SatInt behavior for cost calculations

The saturating arithmetic ensures that cost calculations for extreme values (like dropList with n > Long.MaxValue) properly saturate to Long.MaxValue instead of producing negative values due to overflow.

+++ scalus-core/shared/src/test/scala/scalus/uplc/eval/MemoryUsageTest.scala
@@ -6,23 +6,29 @@ import scalus.uplc.test.ArbitraryInstances
 
 class MemoryUsageTest extends AnyFunSuite with ScalaCheckPropertyChecks with ArbitraryInstances {
     test("memoryUsageInteger(0) == 1") {
-        assert(MemoryUsage.memoryUsageInteger(0) == 1)
-        assert(MemoryUsage.memoryUsageInteger2(0) == 1)
+        assert(MemoryUsage.memoryUsageInteger(0).toLong == 1L)
+        assert(MemoryUsage.memoryUsageInteger2(0).toLong == 1L)
     }
 
     test("memoryUsageInteger(2^64) == 1") {
-        assert(MemoryUsage.memoryUsageInteger(BigInt(2).pow(64) - 1) == 1)
-        assert(MemoryUsage.memoryUsageInteger2(BigInt(2).pow(64) - 1) == 1)
-        assert(MemoryUsage.memoryUsageInteger(BigInt(2).pow(64)) == 2)
-        assert(MemoryUsage.memoryUsageInteger2(BigInt(2).pow(64)) == 2)
-        assert(MemoryUsage.memoryUsageInteger(BigInt(2).pow(64) + 1) == 2)
-        assert(MemoryUsage.memoryUsageInteger2(BigInt(2).pow(64) + 1) == 2)
+        assert(MemoryUsage.memoryUsageInteger(BigInt(2).pow(64) - 1).toLong == 1L)
+        assert(MemoryUsage.memoryUsageInteger2(BigInt(2).pow(64) - 1).toLong == 1L)
+        assert(MemoryUsage.memoryUsageInteger(BigInt(2).pow(64)).toLong == 2L)
+        assert(MemoryUsage.memoryUsageInteger2(BigInt(2).pow(64)).toLong == 2L)
+        assert(MemoryUsage.memoryUsageInteger(BigInt(2).pow(64) + 1).toLong == 2L)
+        assert(MemoryUsage.memoryUsageInteger2(BigInt(2).pow(64) + 1).toLong == 2L)
     }
 
     test("MemoryUsage.memoryUsageInteger(x) == MemoryUsage.memoryUsageInteger2(x)") {
         forAll { (x: BigInt) =>
-            assert(MemoryUsage.memoryUsageInteger(x) == MemoryUsage.memoryUsageInteger2(x))
-            assert(MemoryUsage.memoryUsageInteger(-x) == MemoryUsage.memoryUsageInteger2(-x))
+            assert(
+              MemoryUsage.memoryUsageInteger(x).toLong == MemoryUsage.memoryUsageInteger2(x).toLong
+            )
+            assert(
+              MemoryUsage.memoryUsageInteger(-x).toLong == MemoryUsage
+                  .memoryUsageInteger2(-x)
+                  .toLong
+            )
         }
     }
 }
